moved toggle of collection outside of compose package We still need to pass pack tags instead of only ids
