+++ myExpenses/src/main/java/org/totschnig/myexpenses/compose/State.kt
@@ -33,17 +33,6 @@ fun <K, V> rememberMutableStateMapOf(vararg pairs: Pair<K, V>) : SnapshotStateMa
     }
 }*/
 
-/**
- * true if element was added, false if it was removed
- */
-fun <T> MutableCollection<T>.toggle(element: T) = if (contains(element)) {
-    remove(element)
-    false
-} else {
-    add(element)
-    true
-}
-
 fun <T> MutableState<List<T>>.toggle(element: T) = if (value.contains(element)) {
     value = value - element
     false
