+++ myExpenses/src/main/java/org/totschnig/myexpenses/util/Collection.kt
@@ -1,0 +1,12 @@
+package org.totschnig.myexpenses.util
+
+/**
+ * true if element was added, false if it was removed
+ */
+fun <T> MutableCollection<T>.toggle(element: T) = if (contains(element)) {
+    remove(element)
+    false
+} else {
+    add(element)
+    true
+}
