+++ myExpenses/src/main/java/org/totschnig/myexpenses/viewmodel/MyExpensesViewModel.kt
@@ -29,7 +29,7 @@ import kotlinx.coroutines.withContext
 import org.totschnig.myexpenses.MyApplication
 import org.totschnig.myexpenses.adapter.TransactionPagingSource
 import org.totschnig.myexpenses.compose.ExpansionHandler
-import org.totschnig.myexpenses.compose.toggle
+import org.totschnig.myexpenses.util.toggle
 import org.totschnig.myexpenses.model.*
 import org.totschnig.myexpenses.model.Account
 import org.totschnig.myexpenses.model.Transaction
